requirements.txt
asgiref==3.11.0
django==6.0.1
django-js-asset==3.1.2
django-prose-editor==0.23.1
djangorestframework==3.16.1
nh3==0.3.2
sqlparse==0.5.5
